Moved dashboard.js to pages/index
