+++ src/App.scss
@@ -690,7 +690,7 @@ background-color: #a33a00;
     }
 
     input{
-      background-image:url('./components/dashboard/assets/search.svg');
+      background-image:url('./components/pages/index/assets/search.svg');
       background-position:10px 12px;
       background-repeat: no-repeat;
       background-color: #f7f9f9;
