+++ src/components/pages/index/assets/edge_blur.png
