+++ src/components/pages/index/assets/motorbike_back_wheel.png
