+++ src/components/pages/index/assets/motorbike_background.png
