+++ src/components/pages/index/assets/motorbike_driver.png
