+++ src/components/pages/index/assets/motorbike_front_wheel.png
