+++ src/components/pages/index/assets/motorbike_plant.png
